# openapi.yaml
openapi: 3.0.0
info:
  title: Task Manager
  version: 0.0.1
servers:
  - url: http://localhost:8080/
    description: test task for interview
paths: 
  /Tasks:
    get:
      summary: Метод получения списка задач
      tags:
        - Tasks
      operationId: getAllTasks
      responses:
        '201':
          description: Успешный ответ со списком задач
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tasks"
    post:
      summary: Метод для создания и сохранения новой задачи
      tags:
        - Tasks
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
      responses:
        '201':
          description: Успешное создание задачи

        '400':
          description: Одно из полей введено неверно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskErrorResponse"
  /Tasks/{task_id}:
    delete:
      summary: Удаляем задачу по id
      tags:
        - Tasks
      operationId: deleteTask
      parameters:
        - name: task_id
          in: path
          required: true
          description: Идентификатор задачи
          schema:
            type: string
      responses:
        '200':
          description: Успешное удаление задачи
        '404':
          description: Задачи с переданным айди не существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskErrorResponse"
    post:
      summary: Добавляем комментарии к задаче
      tags:
        - Tasks
        - TaskProperties
      operationId: addComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task_comments"
      responses:
        '200':
          description: Успешное добавление комментария
        '400':
          description: Неправильный фомат комментария
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskErrorResponse"


  /Tasks/{task_id}/{parameter_name}?{parameter_name}={value}:
    patch:
      summary: Метод изменения поля задачи
      tags:
        - TaskProperties
        - Tasks
      operationId: update{property name}
      parameters:
        - name: task_id
          in: path
          required: true
          description: Идентификатор задачи
          schema:
            type: string
        - name: value
          in: path
          required: true
          description: Новое значение параметра
          schema:
            type: string

      responses:
        '200':
          description: Успешное изменение поля
        '400':
          description: Поле введено неверно либо указана несуществующая задача
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyUpdateErrorResponse"

      
        
components:
  schemas:

    TaskErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Заголовок должен быть минимум 2 символа в длину
        timestamp:
          type: string
          format: 2024-11-24T00:00:00+03

    PropertyUpdateErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Заголовок должен быть минимум 2 символа в длину
        timestamp:
          type: string
          format: 2024-11-24T00:00:00+03


    Task_comments:
        type: object
        required:
          - comment
          - task_id
        properties:
          id:
            type: number
            example: 123456
          comment:
            type: string
            example: Новый комментарий
          task_id:
            type: number
            example: 123
    Task:
      type: object
      required:
        - title
        - description
        - status
        - priority
        - author_id
      properties:
        id:
          type: number
          example: 123456789
        title:
          type: string
          example: Покупка хлеба
        description:
          type: string
          example: Нужно взять деньги, пойти в магазин и купить хлеб
        priority:
          type: string
          enum:
            - LOW
            - NORMAL
            - HIGH
            - OVERWORKING
          description: приоритет выполнения задачи
        status:
          type: string
          enum:
            - CREATED
            - IN_PROGRESS
            - COMPLETED
          description: Статус выполнения задачи
        author_id:
          type: integer
          example: 123
        contractor_id:
          type: integer
          example: 123
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Task_comments"
          example:
            id: 1,
            comment: Перед выходом из дома надо надеть шапку
    Tasks:
      type: array
      items:
        $ref: "#/components/schemas/Task"
      